backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
alembic
